Created and passed Interactor Test

+++ src/data_access/FileUserDataAccessObject.java
@@ -101,7 +101,7 @@ public class FileUserDataAccessObject implements SignupUserDataAccessInterface, 
     }
 
     @Override
-    public void editProfile(String username, String name, int age, String gender, double weight, double height) {
+    public void editProfile(String username, String name, int age, String gender, int weight, int height) {
         User user = accounts.get(username);
         user.setName(name);
         user.setAge(age);
